// src/app/shared/models/character/background.ts
export interface Background {
  motivations: {
    traits: string;
    valuedperson: string;
    valuemost: string;
    feelaboutpeople: string;
    valuedpossession: string;
  };
  style: {
    clothes: string;
    hair: string;
    affectations: string;
    ethnicity: string;
    languages: string;
  };
  family?: string;
  siblings?: string;
}